test: fix flaky dashboardLogic jest timeout
