+++ frontend/src/scenes/dashboard/dashboardLogic.test.ts
@@ -24,8 +24,7 @@ import { dashboardsModel } from '~/models/dashboardsModel'
 import { insightsModel } from '~/models/insightsModel'
 import { examples } from '~/queries/examples'
 import { variableDataLogic } from '~/queries/nodes/DataVisualization/Components/Variables/variableDataLogic'
-import { getQueryBasedDashboard } from '~/queries/nodes/InsightViz/utils'
-import { DashboardFilter, HogQLVariable, InsightVizNode, NodeKind, TrendsQuery } from '~/queries/schema/schema-general'
+import { HogQLVariable, InsightVizNode, NodeKind, TrendsQuery } from '~/queries/schema/schema-general'
 import { initKeaTests } from '~/test/init'
 import {
     DashboardPlacement,
@@ -36,30 +35,7 @@ import {
     QueryBasedInsightModel,
 } from '~/types'
 
-import _dashboardJson from './__mocks__/dashboard.json'
-
-const dashboardJson = getQueryBasedDashboard(_dashboardJson as any as DashboardType)!
-
-export function insightOnDashboard(
-    insightId: number,
-    dashboardsRelation: number[],
-    insight: Partial<QueryBasedInsightModel> = {}
-): QueryBasedInsightModel {
-    const tiles = dashboardJson.tiles.filter((tile) => !!tile.insight && tile.insight?.id === insightId)
-    let tile = dashboardJson.tiles[0]
-    if (tiles.length) {
-        tile = tiles[0]
-    }
-    if (!tile.insight) {
-        throw new Error('tile has no insight')
-    }
-    return {
-        ...tile.insight,
-        dashboards: dashboardsRelation,
-        dashboard_tiles: dashboardsRelation.map((dashboardId) => ({ id: insight.id!, dashboard_id: dashboardId })),
-        query: { ...tile.insight.query, ...insight.query, kind: (tile.insight.query?.kind || insight.query?.kind)! },
-    }
-}
+import { dashboardResult, insightOnDashboard, tileFromInsight } from './dashboardLogic.testHelpers'
 
 const TEXT_TILE: DashboardTile<QueryBasedInsightModel> = {
     id: 4,
@@ -80,30 +56,6 @@ const WIDGET_TILE_WITH_CUSTOM_NAME: DashboardTile<QueryBasedInsightModel> = {
     widget: { id: '2', widget_type: 'error_tracking_list', config: {}, name: 'Critical errors' },
     layouts: {},
     color: null,
-}
-
-let tileId = 0
-export const tileFromInsight = (
-    insight: QueryBasedInsightModel,
-    id: number = tileId++
-): DashboardTile<QueryBasedInsightModel> => ({
-    id: id,
-    layouts: {},
-    color: null,
-    insight: insight,
-})
-
-export const dashboardResult = (
-    dashboardId: number,
-    tiles: DashboardTile<QueryBasedInsightModel>[],
-    filters: Partial<DashboardFilter> = {}
-): DashboardType<QueryBasedInsightModel> => {
-    return {
-        ...dashboardJson,
-        filters: { ...dashboardJson.filters, ...filters },
-        id: dashboardId,
-        tiles,
-    }
 }
 
 const uncached = (insight: QueryBasedInsightModel): QueryBasedInsightModel => ({
